use ForceNew on Pipeline team_name
This lets the Terraform SDK handle the process of deleting and
recreating the pipeline resource when the team name changes, with the
added benefit of signalling to the user in the plan that the resource
will be destroyed and recreated. This lets us remove the code in the
provider that handles this special case, too.

Finally, this can work nicely with the `prevent_destroy` lifecycle
meta-argument which will cause Terraform to raise an error if the team
name is changed when `prevent_destroy=true`.

https://developer.hashicorp.com/terraform/plugin/sdkv2/schemas/schema-behaviors#forcenew
https://developer.hashicorp.com/terraform/language/meta-arguments/lifecycle#prevent_destroy